+++ library/src/jvmTest/kotlin/io/github/littlesurvival/jvm/DurationTest.kt
@@ -1,6 +1,5 @@
 package io.github.littlesurvival.jvm
 
-import kotlinx.coroutines.runBlocking
 import kotlin.test.Test
 import kotlin.time.Duration.Companion.hours